feat: comment delete, update

+++ src/main/java/techcourse/myblog/service/ArticleService.java
@@ -3,6 +3,7 @@ package techcourse.myblog.service;
 import org.springframework.stereotype.Service;
 import techcourse.myblog.domain.Article;
 import techcourse.myblog.domain.ArticleRepository;
+import techcourse.myblog.domain.User;
 
 import java.util.Optional;
 
@@ -27,14 +28,21 @@ public class ArticleService {
     }
 
     public boolean tryUpdate(long articleId, Article toUpdate) {
-        return articleRepository.findById(articleId).map(ifExists -> {
-                                                            toUpdate.setId(articleId);
-                                                            articleRepository.save(toUpdate);
-                                                            return true;
+        return articleRepository.findById(articleId).map(original -> {
+                                                            if (original.isSameAuthor(toUpdate)) {
+                                                                toUpdate.setId(articleId);
+                                                                articleRepository.save(toUpdate);
+                                                                return true;
+                                                            }
+                                                            return false;
                                                         }).orElse(false);
     }
 
-    public void delete(long articleId) {
-        articleRepository.deleteById(articleId);
+    public void tryDelete(long articleId, User user) {
+        articleRepository.findById(articleId).ifPresent(article -> {
+            if (article.isSameAuthor(user)) {
+                articleRepository.deleteById(articleId);
+            }
+        });
     }
 }
