+++ src/main/java/techcourse/myblog/service/CommentService.java
@@ -1,8 +1,10 @@
 package techcourse.myblog.service;
 
 import org.springframework.stereotype.Service;
+import techcourse.myblog.domain.Article;
 import techcourse.myblog.domain.Comment;
 import techcourse.myblog.domain.CommentRepository;
+import techcourse.myblog.domain.User;
 
 @Service
 public class CommentService {
@@ -12,15 +14,25 @@ public class CommentService {
         this.commentRepository = commentRepository;
     }
 
-    public boolean tryUpdate(long commentId, Comment toUpdate) {
-        return commentRepository.findById(commentId).map(ifExists -> {
-                                                            toUpdate.setId(commentId);
-                                                            commentRepository.save(toUpdate);
-                                                            return true;
-                                                        }).orElse(false);
+    public Comment write(Article article, User user, String contents) {
+        return commentRepository.save(new Comment(article, user, contents));
     }
 
-    public void delete(long commentId) {
-        commentRepository.deleteById(commentId);
+    public boolean tryUpdate(long commentId, String contents, User author) {
+        return commentRepository.findById(commentId).map(comment -> {
+            if (comment.isSameAuthor(author)) {
+                comment.setContents(contents);
+                return true;
+            }
+            return false;
+        }).orElse(false);
+    }
+
+    public void delete(long commentId, User author) {
+        commentRepository.findById(commentId).ifPresent(comment -> {
+            if (comment.isSameAuthor(author)) {
+                commentRepository.deleteById(commentId);
+            }
+        });
     }
 }
